backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
aiosqlite
alembic
python-jose[cryptography]
bcrypt
python-dotenv
pydantic
pydantic-settings
slowapi
redis
zxcvbn
httpx
